refactor: createInitPlan util 함수들로 분리 (#74)
