+++ src/stores/plan/focusedPlan.ts
@@ -3,6 +3,7 @@ import { create } from 'zustand';
 
 import Plan from '@/plan/Plan';
 import { IPlan } from '@/types/rq/plan';
+import { createInitPlan } from '@/utils/plan/createInitPlan';
 import { changePlanView } from '@/utils/plan/planViewHandlerToMonth';
 
 export type IChangePlanViewType = 'create' | 'edit' | null;
@@ -41,29 +42,6 @@ const initialState = {
   focusedPlan: null,
   isDragging: false,
 } as const;
-
-const createInitPlan = (planData: Partial<IPlan>) => {
-  const newPlan = new Plan({
-    id: -1,
-    title: '새로운 일정',
-    description: null,
-    isAllDay: false,
-    type: 'task',
-    color: '#52D681',
-    categoryId: null,
-    tags: [],
-    startTime: '',
-    endTime: '',
-    ...planData,
-  });
-  if (!planData.startTime) {
-    newPlan._startTime = Date.now();
-  }
-  if (!planData.endTime) {
-    newPlan._endTime = Date.now() + 1000 * 60 * 30;
-  }
-  return newPlan;
-};
 
 const useFocusedPlanState = create<IFocusedPlanState & IFocusedPlanAction>(
   (set, get) => ({
